+++ 42Bangkok_L/Rush02/ex00_My_Version/rsh02.h
@@ -6,7 +6,7 @@
 /*   By: kmahanin <kmahanin@student.42bangkok.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/24 21:46:42 by kmahanin          #+#    #+#             */
-/*   Updated: 2026/07/27 05:54:12 by kmahanin         ###   ########.fr       */
+/*   Updated: 2026/07/27 14:39:15 by kmahanin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@
 
 typedef struct s_dict
 {
-	unsigned long long	nb;
+	unsigned __int128	nb;
 	char				*val;
 }						t_dict;
 
 int						input_validation(int argc, char **argv);
 void					putstr(char *str);
 int						len(char *str);
-unsigned long long		atoull(char *str);
+unsigned __int128		atoull(char *str);
 char					*ft_strdup(char *src);
 char					*get_dict_contact(char *path);
 
@@ -34,6 +34,6 @@ t_dict					*parse_dict(char *path);
 void					free_dict(t_dict *dict);
 void					sort_dict(t_dict *dict, int size);
 
-void					convert_number(unsigned long long n, t_dict *dict,
+void					convert_number(unsigned __int128 n, t_dict *dict,
 							int *first);
 #endif
